Move shareable fillSearchable util to new file
